refactored cloud funciton
