+++ devsecops/artifact-registry-vulnerability-scanning/cloud-function/src/save-to-bucket.js
@@ -1,0 +1,17 @@
+const { Storage } = require('@google-cloud/storage');
+const storage = new Storage();
+
+const writeVulnToBucket = async (bucketName, vulnText, destinationObjectName) => {
+  try {
+    const bucket = storage.bucket(bucketName);
+    const file = bucket.file(destinationObjectName);
+    
+    await file.save(vulnText);
+    console.log(`Saved ${destinationObjectName} successfully to ${bucketName} bucket`);
+  } catch (error) {
+    console.error(`Failed to write ${destinationObjectName} to bucket: ${error}`);
+    throw error;
+  }
+};
+
+module.exports = { writeVulnToBucket };
